+++ SpaceService/SpaceService/Model/PlayerState.cs
@@ -10,14 +10,14 @@ namespace SpaceService.Model
         public Player Player { get; set; }
         public Vector Position { get; set; }
         public bool Finished { get; set; }
-        public float Score { get; set; }
+        public int Score { get; set; }
         public bool ResultRequested { get; set; }
 
         public PlayerState()
         {
             Position = new Vector();
             Finished = false;
-            Score = float.MaxValue;
+            Score = int.MaxValue;
             ResultRequested = false;
         }
     }
